Dungeon Example
Added a lightweight scene consisting of modular assets, creating a dungeon to be used as an example.
