+++ assets/materials/floor.mat
@@ -1,0 +1,4 @@
+(
+    friendly_name: "Floor",
+    base_color: Some((1.0, 1.0, 1.0, 1.0)),
+)
