+++ assets/models/0_25_deep_Floor.obj
@@ -1,0 +1,35 @@
+# Blender 4.5.2 LTS
+# www.blender.org
+mtllib 0_25_deep_Floor.mtl
+o 0.25_deep_floor
+v 0.000000 0.000000 1.000000
+v 1.000000 0.000000 1.000000
+v 0.000000 0.000000 0.000000
+v 1.000000 0.000000 -0.000000
+v 0.000000 -0.125000 0.000000
+v 0.000000 -0.125000 1.000000
+v 1.000000 -0.125000 1.000000
+v 1.000000 -0.125000 -0.000000
+vn -0.0000 1.0000 -0.0000
+vn -0.0000 -1.0000 -0.0000
+vn -0.0000 -0.0000 -1.0000
+vn -0.0000 -0.0000 1.0000
+vn 1.0000 -0.0000 -0.0000
+vn -1.0000 -0.0000 -0.0000
+vt 1.000000 0.000000
+vt 0.000000 1.000000
+vt 0.000000 0.000000
+vt 1.000000 1.000000
+s 0
+f 2/1/1 3/2/1 1/3/1
+f 5/2/2 7/1/2 6/3/2
+f 8/4/3 3/2/3 4/4/3
+f 6/3/4 2/1/4 1/3/4
+f 7/1/5 4/4/5 2/1/5
+f 5/2/6 1/3/6 3/2/6
+f 2/1/1 4/4/1 3/2/1
+f 5/2/2 8/4/2 7/1/2
+f 8/4/3 5/2/3 3/2/3
+f 6/3/4 7/1/4 2/1/4
+f 7/1/5 8/4/5 4/4/5
+f 5/2/6 6/3/6 1/3/6
