+++ assets/models/0_25m_Trim.obj
@@ -1,0 +1,116 @@
+# Blender 4.5.2 LTS
+# www.blender.org
+mtllib 0_25m_Trim.mtl
+o Cube.004
+v 0.666666 1.250000 -0.000000
+v -0.000000 1.000000 -0.000000
+v -0.000000 1.250000 -0.000000
+v 0.333333 1.250000 -0.000000
+v 0.333333 1.000000 -0.000000
+v 1.000000 1.250000 -0.000000
+v 1.000000 1.000000 -0.000000
+v 0.666667 1.000000 -0.000000
+v 0.313333 1.020000 0.020000
+v 0.313333 1.230000 0.020000
+v 0.020000 1.230000 0.020000
+v 0.020000 1.020000 0.020000
+v 0.980000 1.020000 0.020000
+v 0.980000 1.230000 0.020000
+v 0.686666 1.230000 0.020000
+v 0.686666 1.020000 0.020000
+v 0.646666 1.020000 0.020000
+v 0.646667 1.230000 0.020000
+v 0.353333 1.230000 0.020000
+v 0.353333 1.020000 0.020000
+v 0.666666 1.250000 -0.072847
+v -0.000000 1.000000 -0.072847
+v -0.000000 1.250000 -0.072847
+v 0.333333 1.250000 -0.072847
+v 0.333333 1.000000 -0.072846
+v 1.000000 1.250000 -0.072847
+v 1.000000 1.000000 -0.072846
+v 0.666667 1.000000 -0.072846
+vn -0.0000 -0.0000 1.0000
+vn 0.7071 -0.0000 0.7071
+vn -0.0000 0.7071 0.7071
+vn -0.7071 -0.0000 0.7071
+vn -0.0000 -0.7071 0.7071
+vn -0.0000 -1.0000 -0.0000
+vn -0.0000 1.0000 -0.0000
+vn 1.0000 -0.0000 -0.0000
+vn -1.0000 -0.0000 -0.0000
+vn -0.0000 -0.0000 -1.0000
+vt 0.567500 0.835833
+vt 0.620000 0.872500
+vt 0.567500 0.872500
+vt 0.567500 0.752500
+vt 0.620000 0.789167
+vt 0.567500 0.789167
+vt 0.567500 0.794167
+vt 0.620000 0.830833
+vt 0.567500 0.830833
+vt 0.625000 0.833333
+vt 0.562500 0.833333
+vt 0.620000 0.835833
+vt 0.625000 0.875000
+vt 0.562500 0.875000
+vt 0.562500 0.750000
+vt 0.620000 0.752500
+vt 0.625000 0.750000
+vt 0.562500 0.791667
+vt 0.625000 0.791667
+vt 0.620000 0.794167
+s 0
+usemtl Material
+f 9/1/1 11/2/1 12/3/1
+f 13/4/1 15/5/1 16/6/1
+f 17/7/1 19/8/1 20/9/1
+f 4/10/2 9/1/2 5/11/2
+f 4/10/3 11/2/3 10/12/3
+f 3/13/4 12/3/4 11/2/4
+f 5/11/5 12/3/5 2/14/5
+f 7/15/2 14/16/2 13/4/2
+f 6/17/3 15/5/3 14/16/3
+f 8/18/4 15/5/4 1/19/4
+f 8/18/5 13/4/5 16/6/5
+f 8/18/2 18/20/2 17/7/2
+f 1/19/3 19/8/3 18/20/3
+f 5/11/4 19/8/4 4/10/4
+f 5/11/5 17/7/5 20/9/5
+f 8/18/6 27/15/6 7/15/6
+f 1/19/7 24/10/7 4/10/7
+f 6/17/8 27/15/8 26/17/8
+f 2/14/6 25/11/6 5/11/6
+f 3/13/9 22/14/9 2/14/9
+f 4/10/7 23/13/7 3/13/7
+f 5/11/6 28/18/6 8/18/6
+f 6/17/7 21/19/7 1/19/7
+f 28/18/10 26/17/10 27/15/10
+f 25/11/10 21/19/10 28/18/10
+f 22/14/10 24/10/10 25/11/10
+f 9/1/1 10/12/1 11/2/1
+f 13/4/1 14/16/1 15/5/1
+f 17/7/1 18/20/1 19/8/1
+f 4/10/2 10/12/2 9/1/2
+f 4/10/3 3/13/3 11/2/3
+f 3/13/4 2/14/4 12/3/4
+f 5/11/5 9/1/5 12/3/5
+f 7/15/2 6/17/2 14/16/2
+f 6/17/3 1/19/3 15/5/3
+f 8/18/4 16/6/4 15/5/4
+f 8/18/5 7/15/5 13/4/5
+f 8/18/2 1/19/2 18/20/2
+f 1/19/3 4/10/3 19/8/3
+f 5/11/4 20/9/4 19/8/4
+f 5/11/5 8/18/5 17/7/5
+f 8/18/6 28/18/6 27/15/6
+f 1/19/7 21/19/7 24/10/7
+f 6/17/8 7/15/8 27/15/8
+f 2/14/6 22/14/6 25/11/6
+f 3/13/9 23/13/9 22/14/9
+f 4/10/7 24/10/7 23/13/7
+f 5/11/6 25/11/6 28/18/6
+f 6/17/7 26/17/7 21/19/7
+f 28/18/10 21/19/10 26/17/10
+f 25/11/10 24/10/10 21/19/10
+f 22/14/10 23/13/10 24/10/10
