+++ assets/models/1m_Floor.mtl
@@ -1,0 +1,2 @@
+# Blender 4.5.2 LTS MTL File: 'bevy deafult scene.blend'
+# www.blender.org
